% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchFunction.r
\name{Search}
\alias{Search}
\title{Search via RSQLite (equality).}
\usage{
Search(conn, table_name, column_names, values)
}
\arguments{
\item{conn}{Object of Class SQLiteConnection connected to the database in
which the table is present.}

\item{table_name}{String.}

\item{column_names}{Vector of column names on which the conditions are to be
applied.}

\item{values}{Vector of conditional values in the same order as column_names.}
}
\value{
Result of the search query in a dataframe format.
}
\description{
This Function Searches the database for values given by the user and returns
the result.It asks the user for the columns and their corresponding values
and then performs an equality search query (like: "SELECT * FROM TABLE1 WHERE
COLUMN1 = VALUE1") on the database Connection provided.
}
\examples{
\code{
conn <- dbConnect (RSQLite::SQLite(), "mtcars.db")
tables <- dbListTables(conn)
ans <- Search(conn, tables[1], c('disp', 'wt'), c(160, 2.620))
}

}
